Move Sleeper mapping helper to shared module
